+++ frontend/src/locales/en.json
@@ -28,5 +28,7 @@
   "avatarImagePreview": "Avatar Image Preview",
   "enterEmailAddress": "Enter your email address",
   "language": "Language",
-  "save": "Save"
+  "save": "Save",
+  "leaderboard": "Leaderboard",
+  "backToHome": "Back to Home"
 }
