+++ frontend/src/pages/Leaderboard/index.ts
@@ -1,12 +1,21 @@
 import { Page } from '@/core/Page';
 import CommonLayout from '@/layouts/common/index';
 import { checkUserAccess } from '@/models/User/auth';
-import { fetchUsers } from '@/models/User/repository';
+import { fetchCurrentUser, fetchUsers, IUserData } from '@/models/User/repository';
+import i18next from 'i18next';
 
 interface IRankingUser {
   username: string;
   level: number;
 }
+
+const updateContent = () => {
+  const titleEl = document.querySelector('.ranking-container h1');
+  if (titleEl) titleEl.textContent = i18next.t('leaderboard');
+
+  const saveButton = document.querySelector('.btn');
+  if (saveButton) saveButton.textContent = i18next.t('backToHome');
+};
 
 const LeaderboardPage = new Page({
   name: 'Leaderboard',
@@ -19,6 +28,12 @@ const LeaderboardPage = new Page({
     const backBtn = document.getElementById('backBtn');
 
     try {
+      const userData: IUserData = await fetchCurrentUser();
+      if (userData.language) {
+        document.documentElement.lang = userData.language;
+        i18next.changeLanguage(userData.language, updateContent);
+      }
+
       const users: IRankingUser[] = await fetchUsers();
       const sortedUsers = users.sort((a, b) => b.level - a.level).slice(0, 10);
